added NPM+R

+++ resources/scripts/start.sh
@@ -2,7 +2,7 @@
 source $(dirname "$0")/funcs/internal.sh
 source $(dirname "$0")/funcs/environment.sh
 
-goan_version="GOAN v1.0.9"
+goan_version="GOAN v1.1.0"
 goan_log_path="/opt/log"
 
 echo -e "\n${goan_version}\n"
@@ -15,37 +15,64 @@ if [[ ! -d "/var/www/html" ]]; then
     mkdir /var/www/html
 fi
 
+# CLEAN UP
+goan_config="/goaccess-config/goaccess.conf"
+cp /goaccess-config/goaccess.conf.bak ${goan_config}
+
+nginx_html="/var/www/html/index.html"
+if [[ -f ${nginx_html} ]]; then
+    rm ${nginx_html}
+fi
+
+html_config="/var/www/html/goaccess_conf.html"
+if [[ -f ${html_config} ]]; then
+    rm ${html_config}
+fi
+# END OF CLEAN UP
+
 # BEGIN PROXY LOGS
-goan_proxy_config="/goaccess-config/goaccess_proxy.conf"
-cp /goaccess-config/goaccess.conf ${goan_proxy_config}
+if [[ -z "${LOG_TYPE}" || "${LOG_TYPE}" == "NPM" || "${LOG_TYPE}" == "NPM+R"]]
+then
+    goan_proxy_archive_log="/goaccess-config/access_archive.log"
+    if [[ -f ${goan_proxy_archive_log} ]]; then
+        rm ${goan_proxy_archive_log}
+    fi
 
-nginx_proxy_html="/var/www/html/index.html"
-if [[ -f ${nginx_proxy_html} ]]; then
-    rm ${nginx_proxy_html}
+    set_npm_proxy_config ${goan_config}
+    load_proxy_logs ${goan_log_path} ${goan_config} ${goan_proxy_archive_log} ${nginx_html}
+
+    if [[ "${LOG_TYPE}" == "NPM+R" ]]; then
+        #redirection logging
+        goan_redirection_config="/goaccess-config/goaccess_redirection.conf"
+        cp /goaccess-config/goaccess.conf.bak ${goan_redirection_config}
+
+        set_npm_redirection_config ${goan_redirection_config}
+        load_proxy_logs ${goan_log_path} ${goan_redirection_config} ${goan_proxy_archive_log} ${nginx_html}
+   fi
+elif [[ "${LOG_TYPE}" == "TRAEFIK" ]]
+then
+    set_traefik_config ${goan_config}
+    load_traefik_logs ${goan_log_path} ${goan_config} ${nginx_html}
 fi
-
-proxy_html_config="/var/www/html/goaccess_conf.html"
-if [[ -f ${proxy_html_config} ]]; then
-    rm ${proxy_html_config}
-fi
-
-goan_proxy_archive_log="/goaccess-config/access_archive.log"
-if [[ -f ${goan_proxy_archive_log} ]]; then
-    rm ${goan_proxy_archive_log}
-fi
-
-set_proxy_config ${goan_proxy_config}
-load_proxy_logs ${goan_log_path} ${goan_proxy_config} ${goan_proxy_archive_log} ${nginx_proxy_html}
-exclude_ips ${goan_proxy_config}
-debug ${proxy_html_config} ${goan_proxy_config}
 # END PROXY LOGS
+
+# BEGIN GLOBAL SETTINGS
+exclude_ips             ${goan_config}
+debug                   ${goan_config} ${html_config}
+set_geoip_database      ${goan_config}
+# END GLOBAL SETTINGS
 
 #RUN NGINX
 tini -s -- nginx
 
 #RUN GOACCESS
 echo -e "\nRUN MAIN GOACCESS"
-tini -s -- /goaccess/goaccess --daemonize --no-global-config --config-file=${goan_proxy_config}
+tini -s -- /goaccess/goaccess --daemonize --no-global-config --config-file=${goan_config}
+
+#RUN SECOND GOACCESS
+if [[ "${LOG_TYPE}" == "NPM+R" ]]; then
+    tini -s -- /goaccess/goaccess --daemonize --no-global-config --config-file=${goan_redirection_config}
+fi
 
 #Leave container running
 tail -f /dev/null
